requirements.txt
aiofiles==0.8.0
async-timeout==5.0.1
androidstorage4kivy==0.1.1
briefcase==0.3.22
cffi==1.17.1
eyed3==0.9.7
hmmlearn==0.3.3
imblearn==0.0
kivy==2.3.1
librosa==0.10.2.post1
matplotlib>=3.8.3
numpy==1.24.3
python-for-android==2024.1.21
pyaudio==0.2.14
pyAudioAnalysis==0.3.14
pytest==8.3.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
pytest-cov>=3.0
pyjnius==1.6.1
pytest-timeout>=2.1.0
python-dotenv==1.0.1
psutil==6.1.1
pydub==0.25.1
plotly==6.0.0
plyer==2.1.0
regex==2024.11.6
requests==2.32.3
soundfile==0.13.1
static-ffmpeg==2.8
scipy==1.15.2
sounddevice==0.5.1
scikit-learn==1.6.1
toga==0.4.9
toga_core==0.4.9
tenacity==9.0.0
vosk==0.3.45
pywin32>=2.2.8